+++ nodalia-cards-2.1.2-alpha.1.js
@@ -1,2 +1,0 @@
-import "./nodalia-cards-2.1.2-alpha.2.js";
-if(typeof window!=="undefined"){window.__NODALIA_LOADER__={"mode":"compat","pkgVersion":"2.1.2-alpha.2","contentSha256_12":"ba544bc3e20a","file":"nodalia-cards-2.1.2-alpha.1.js","targetFile":"nodalia-cards-2.1.2-alpha.2.js","fallbackFile":"nodalia-cards.js"};}