+++ nodalia-cards.manifest.js
@@ -1,23 +1,23 @@
 export default {
-  "pkgVersion": "2.1.2-alpha.2",
-  "contentSha256_12": "ba544bc3e20a",
+  "pkgVersion": "2.1.2-alpha.3",
+  "contentSha256_12": "ba61578a77c1",
   "file": "nodalia-cards.bundle.js",
   "loaderFile": "nodalia-cards.js",
   "hacsFile": "nodalia-cards.js",
   "compatLoaderFiles": [
     "nodalia-cards-2.1.1.js",
-    "nodalia-cards-2.1.2-alpha.1.js"
+    "nodalia-cards-2.1.2-alpha.2.js"
   ],
-  "splitCoreFile": "nodalia-cards-core-2.1.2-alpha.2.js",
+  "splitCoreFile": "nodalia-cards-core-2.1.2-alpha.3.js",
   "splitCoreSha256_12": "fb9704f55dc6",
-  "splitSuiteFile": "nodalia-cards-suite-2.1.2-alpha.2.js",
-  "splitSuiteSha256_12": "7f7f5bcc2352",
-  "editorFile": "nodalia-cards-editor-2.1.2-alpha.2.js",
+  "splitSuiteFile": "nodalia-cards-suite-2.1.2-alpha.3.js",
+  "splitSuiteSha256_12": "522029f93187",
+  "editorFile": "nodalia-cards-editor-2.1.2-alpha.3.js",
   "editorSha256_12": "7f50f338e82a"
 };
-export const pkgVersion = "2.1.2-alpha.2";
-export const contentSha256_12 = "ba544bc3e20a";
+export const pkgVersion = "2.1.2-alpha.3";
+export const contentSha256_12 = "ba61578a77c1";
 export const file = "nodalia-cards.bundle.js";
-export const splitCoreFile = "nodalia-cards-core-2.1.2-alpha.2.js";
-export const splitSuiteFile = "nodalia-cards-suite-2.1.2-alpha.2.js";
-export const editorFile = "nodalia-cards-editor-2.1.2-alpha.2.js";
+export const splitCoreFile = "nodalia-cards-core-2.1.2-alpha.3.js";
+export const splitSuiteFile = "nodalia-cards-suite-2.1.2-alpha.3.js";
+export const editorFile = "nodalia-cards-editor-2.1.2-alpha.3.js";